refactor: 패키지 구조 변경
- request 하위에 도메인 별로 분류
