+++ src/main/java/jpabook/dashdine/dto/request/user/EmailAuthRequestDto.java
@@ -1,10 +1,11 @@
-package jpabook.dashdine.dto.request;
+package jpabook.dashdine.dto.request.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class EmailRequestDto {
+public class EmailAuthRequestDto {
+    private String loginId;
     private String email;
 }
