+++ src/main/java/jpabook/dashdine/dto/request/user/EmailRequestDto.java
@@ -1,11 +1,10 @@
-package jpabook.dashdine.dto.request;
+package jpabook.dashdine.dto.request.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class EmailAuthRequestDto {
-    private String loginId;
+public class EmailRequestDto {
     private String email;
 }
