+++ src/main/java/jpabook/dashdine/service/user/UserManagementService.java
@@ -4,8 +4,8 @@ import jpabook.dashdine.domain.user.PasswordManager;
 import jpabook.dashdine.domain.user.User;
 import jpabook.dashdine.domain.user.UserRoleEnum;
 import jpabook.dashdine.dto.request.user.DeactivateRequestDto;
-import jpabook.dashdine.dto.request.PasswordChangeRequestDto;
-import jpabook.dashdine.dto.request.SignupRequestDto;
+import jpabook.dashdine.dto.request.user.PasswordChangeRequestDto;
+import jpabook.dashdine.dto.request.user.SignupRequestDto;
 import jpabook.dashdine.redis.RedisUtil;
 import jpabook.dashdine.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
